+++ database/migrations/2022_03_21_092204_create_geofences_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->float("longitude");
             $table->float("latitude");
-            $table->double("radius");
+            $table->float("radius");
             $table->timestamps();
         });
     }
